+++ app/src/main/java/org/oxycblt/auxio/detail/AlbumDetailFragment.kt
@@ -42,6 +42,7 @@ import org.oxycblt.auxio.settings.Settings
 import org.oxycblt.auxio.ui.Header
 import org.oxycblt.auxio.ui.Item
 import org.oxycblt.auxio.ui.MenuFragment
+import org.oxycblt.auxio.ui.Sort
 import org.oxycblt.auxio.util.applySpans
 import org.oxycblt.auxio.util.canScroll
 import org.oxycblt.auxio.util.collect
@@ -143,11 +144,16 @@ class AlbumDetailFragment :
     override fun onShowSortMenu(anchor: View) {
         menu(anchor, R.menu.menu_album_sort) {
             val sort = detailModel.albumSort
-            requireNotNull(menu.findItem(sort.itemId)).isChecked = true
+            requireNotNull(menu.findItem(sort.mode.itemId)).isChecked = true
             requireNotNull(menu.findItem(R.id.option_sort_asc)).isChecked = sort.isAscending
             setOnMenuItemClickListener { item ->
                 item.isChecked = !item.isChecked
-                detailModel.albumSort = requireNotNull(sort.assignId(item.itemId))
+                detailModel.albumSort =
+                    if (item.itemId == R.id.option_sort_asc) {
+                        sort.withAscending(item.isChecked)
+                    } else {
+                        sort.withMode(requireNotNull(Sort.Mode.fromItemId(item.itemId)))
+                    }
                 true
             }
         }