+++ app/src/main/java/org/oxycblt/auxio/detail/GenreDetailFragment.kt
@@ -41,6 +41,7 @@ import org.oxycblt.auxio.settings.Settings
 import org.oxycblt.auxio.ui.Header
 import org.oxycblt.auxio.ui.Item
 import org.oxycblt.auxio.ui.MenuFragment
+import org.oxycblt.auxio.ui.Sort
 import org.oxycblt.auxio.util.applySpans
 import org.oxycblt.auxio.util.collect
 import org.oxycblt.auxio.util.collectImmediately
@@ -139,11 +140,16 @@ class GenreDetailFragment :
     override fun onShowSortMenu(anchor: View) {
         menu(anchor, R.menu.menu_genre_sort) {
             val sort = detailModel.genreSort
-            requireNotNull(menu.findItem(sort.itemId)).isChecked = true
+            requireNotNull(menu.findItem(sort.mode.itemId)).isChecked = true
             requireNotNull(menu.findItem(R.id.option_sort_asc)).isChecked = sort.isAscending
             setOnMenuItemClickListener { item ->
                 item.isChecked = !item.isChecked
-                detailModel.genreSort = requireNotNull(sort.assignId(item.itemId))
+                detailModel.genreSort =
+                    if (item.itemId == R.id.option_sort_asc) {
+                        sort.withAscending(item.isChecked)
+                    } else {
+                        sort.withMode(requireNotNull(Sort.Mode.fromItemId(item.itemId)))
+                    }
                 true
             }
         }